broke out firebaseFirestore into individual files
