+++ src/index.js
@@ -1,8 +1,13 @@
 import './styles/styles.scss';
 import { setDate } from './components/setDate';
 import { asideExpand } from './components/asideExpand';
-import { newChore } from './components/newChore';
 import { firebaseAuth } from './components/firebase/firebaseAuth';
 import { firebaseFirestore } from './components/firebase/firebaseFirestore';
+// import { completeChore } from './components/firebase/firestore/completeChore';
 
-firebaseAuth.then(firebaseFirestore);
+firebaseAuth.then(firebaseFirestore); 
+
+const completeChore = () => {
+    const parent = this.parentElement;
+	console.log(parent);
+}
