+++ src/styles/components/editChoreModal.scss
@@ -73,3 +73,39 @@
     }
 
 }
+
+#confirm-delete-modal {
+    height: 100%;
+	width: 100%;
+	position: absolute;
+	top: 0;
+	left: 0;
+	display: grid;
+	align-content: center;
+	justify-content: center;
+    background-color: rgba(0,0,0,0.2);
+    z-index: 200;
+
+    &.hide {
+        display: none;
+    }
+
+    &>div {
+        background-color: $lt-gray;
+        padding: 50px;
+        display: grid;
+        grid-template-columns: repeat(2, 1fr);
+        grid-column-gap: 50px;
+
+        h2 {
+            grid-column: 1 / 3;
+            margin-bottom: 20px;
+        }
+
+        #confirm-delete {
+            background-color: $red;
+        }
+
+    }
+
+}
